Instance weights were being cast to ints for some reason. Removed the casts
git-svn-id: https://svn.cms.waikato.ac.nz/svn/weka/trunk/weka@4098 e0a1b77d-ad91-4216-81b1-defd5f83fa92

+++ classifiers/bayes/AODE.java
@@ -96,7 +96,7 @@ import java.util.Vector;
  *
  * @author Janice Boughton (jrbought@csse.monash.edu.au)
  * @author Zhihai Wang (zhw@csse.monash.edu.au)
- * @version $Revision: 1.17 $
+ * @version $Revision: 1.18 $
  */
 public class AODE extends Classifier
     implements OptionHandler, WeightedInstancesHandler, UpdateableClassifier, 
@@ -320,7 +320,7 @@ public class AODE extends Classifier
        return;   // ignore instances with missing class
 
     int classVal = (int)instance.classValue();
-    int weight = (int)instance.weight();
+    double weight = instance.weight();
  
     m_ClassCounts[classVal] += weight;
     m_SumInstances += weight;
